requirements.txt
google-genai>=1.16.0
openpyxl>=3.1.0
pypdf>=4.2.0